requirements.txt
asgiref==3.4.1
braintree==4.12.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==0.1
cryptography==3.4.8
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-payments==0.14.0
django-translations==1.3.0
idna==3.2
Pillow==8.3.1
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
suds-jurko==0.6
urllib3==1.26.6
xmltodict==0.12.0
